- Fixed window resizing not working properly on Ventura - Updated code formats - Modified overlay to be a lot more reliable
