+++ SwiftAA/ViewModels/Windows/Main/DataHandler.swift
